Update aria-label in Select.vue component
The aria-label in the Select.vue component has been modified to be dynamic rather than have a fixed value. A new property 'ariaLabel' has been introduced to set the value of 'aria-label' for the search input. Now, the default value is 'Search for option', but it can be customized as needed.